Struct graphviz::Id

pub struct Id<'a> {
    // some fields omitted
}

Unstable (rustc_private)

Id is a Graphviz ID.

Methods

impl<'a> Id<'a>

fn new<Name: IntoCow<'a, str>>(name: Name) -> Result<Id<'a>, ()>

Unstable (rustc_private)

Creates an Id named name.

The caller must ensure that the input conforms to an identifier format: it must be a non-empty string made up of alphanumeric or underscore characters, not beginning with a digit (i.e. the regular expression [a-zA-Z_][a-zA-Z_0-9]*).

(Note: this format is a strict subset of the ID format defined by the DOT language. This function may change in the future to accept a broader subset, or the entirety, of DOT's ID format.)

Passing an invalid string (containing spaces, brackets, quotes, ...) will return an empty Err value.

fn as_slice(&'a self) -> &'a str

Unstable (rustc_private)

fn name(self) -> Cow<'a, str>

Unstable (rustc_private)