Trait graphviz::GraphWalk

pub trait GraphWalk<'a> {
    type Node: Clone;
    type Edge: Clone;
    fn nodes(&'a self) -> Nodes<'a, Self::Node>;

    fn edges(&'a self) -> Edges<'a, Self::Edge>;

    fn source(&'a self, edge: &Self::Edge) -> Self::Node;

    fn target(&'a self, edge: &Self::Edge) -> Self::Node;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

GraphWalk is an abstraction over a directed graph = (nodes,edges) made up of node handles N and edge handles E, where each E can be mapped to its source and target nodes.

The lifetime parameter 'a is exposed in this trait (rather than introduced as a generic parameter on each method declaration) so that a client impl can choose N and E that have substructure that is bound by the self lifetime 'a.

The nodes and edges method each return instantiations of Cow<[T]> to leave implementors the freedom to create entirely new vectors or to pass back slices into internally owned vectors.

Associated Types

type Node: Clone

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type Edge: Clone

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn nodes(&'a self) -> Nodes<'a, Self::Node>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns all the nodes in this graph.

fn edges(&'a self) -> Edges<'a, Self::Edge>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns all of the edges in this graph.

fn source(&'a self, edge: &Self::Edge) -> Self::Node

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The source node for edge.

fn target(&'a self, edge: &Self::Edge) -> Self::Node

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The target node for edge.