Trait graphviz::IntoCow

pub trait IntoCow<'a, B: ?Sized> where
    B: ToOwned, {
    fn into_cow(self) -> Cow<'a, B>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn into_cow(self) -> Cow<'a, B>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?