Enum graphviz::LabelText

pub enum LabelText<'a> {
    LabelStr(Cow<'a, str>),
    EscStr(Cow<'a, str>),
    HtmlStr(Cow<'a, str>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The text for a graphviz label on a node or edge.

Variants
LabelStr(Cow<'a, str>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This kind of label preserves the text directly as is.

Occurrences of backslashes (\) are escaped, and thus appear as backslashes in the rendered label.

EscStr(Cow<'a, str>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This kind of label uses the graphviz label escString type: http://www.graphviz.org/content/attrs#kescString

Occurrences of backslashes (\) are not escaped; instead they are interpreted as initiating an escString escape sequence.

Escape sequences of particular interest: in addition to \n to break a line (centering the line preceding the \n), there are also the escape sequences \l which left-justifies the preceding line and \r which right-justifies it.

HtmlStr(Cow<'a, str>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This uses a graphviz HTML string label. The string is printed exactly as given, but between < and >. No escaping is performed.

Methods

impl<'a> LabelText<'a>

pub fn label<S: IntoCow<'a, str>>(s: S) -> LabelText<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn escaped<S: IntoCow<'a, str>>(s: S) -> LabelText<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn html<S: IntoCow<'a, str>>(s: S) -> LabelText<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn to_dot_string(&self) -> String

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Renders text as string suitable for a label in a .dot file. This includes quotes or suitable delimiters.

pub fn prefix_line(self, prefix: LabelText) -> LabelText<'static>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Puts prefix on a line above this label, with a blank line separator.

pub fn suffix_line(self, suffix: LabelText) -> LabelText<'static>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Puts suffix on a line below this label, with a blank line separator.