Enum graphviz::LabelTextUnstable

pub enum LabelText<'a> {
    LabelStr(CowString<'a>),
    EscStr(CowString<'a>),
}

The text for a graphviz label on a node or edge.

Variants

LabelStr

This kind of label preserves the text directly as is.

Occurrences of backslashes (\) are escaped, and thus appear as backslashes in the rendered label.

EscStr

This kind of label uses the graphviz label escString type: http://www.graphviz.org/content/attrs#kescString

Occurrences of backslashes (\) are not escaped; instead they are interpreted as initiating an escString escape sequence.

Escape sequences of particular interest: in addition to \n to break a line (centering the line preceding the \n), there are also the escape sequences \l which left-justifies the preceding line and \r which right-justifies it.

Methods

impl<'a> LabelText<'a>

fn label<S: IntoCow<'a, String, str>>(s: S) -> LabelText<'a>

fn escaped<S: IntoCow<'a, String, str>>(s: S) -> LabelText<'a>

fn escape(&self) -> String

Renders text as string suitable for a label in a .dot file.

fn prefix_line(self, prefix: LabelText) -> LabelText<'static>

Puts prefix on a line above this label, with a blank line separator.

fn suffix_line(self, suffix: LabelText) -> LabelText<'static>

Puts suffix on a line below this label, with a blank line separator.